+++ tests/unit/test_harmonic_model.py
@@ -214,6 +214,37 @@ def test_fit_harmonic_model_checks():
     with pytest.raises(ValueError, match="Monthly target data must start with January"):
         mesmer.stats.fit_harmonic_model(yearly_predictor, monthly_target)
 
+    monthly_target["time"] = pd.date_range("2000-01-01", periods=10 * 12, freq="ME")
+
+    with pytest.raises(ValueError, match="yearly_predictor should be 2D, but is 1D"):
+        mesmer.stats.fit_harmonic_model(yearly_predictor.isel(cells=0), monthly_target)
+
+    with pytest.raises(ValueError, match="Differently named `gridcell_dim`:"):
+        mesmer.stats.fit_harmonic_model(
+            yearly_predictor.rename(cells="gp"), monthly_target
+        )
+
+    with pytest.raises(ValueError, match="Differently named `gridcell_dim`:"):
+        mesmer.stats.fit_harmonic_model(
+            yearly_predictor, monthly_target.rename(cells="gp")
+        )
+
+    with pytest.raises(
+        ValueError,
+        match=r"yearly_predictor` and `monthly_target` don't have the same number of gridcells \(6 vs\. 4\)",
+    ):
+        mesmer.stats.fit_harmonic_model(
+            yearly_predictor, monthly_target.isel(cells=slice(None, 4))
+        )
+
+    with pytest.raises(
+        ValueError,
+        match=r"yearly_predictor` and `monthly_target` don't have the same number of gridcells \(5 vs\. 6\)",
+    ):
+        mesmer.stats.fit_harmonic_model(
+            yearly_predictor.isel(cells=slice(None, 5)), monthly_target
+        )
+
 
 def test_fit_harmonic_model_time_dim():
     # test if the time dimension can be different from "time"